refactor(rooch-da): centralize backend management logic
Extract backend management logic to a shared `DABackends` struct. This removes duplication and simplifies backend initialization in both `mod.rs` and `server.rs`.
